Moved everything from `Inachis\Component\Common` namespace into `Inachis\Component\CoreBundle` namespace for consistency
