+++ src/Inachis/CoreBundle/Exception/FileNotFoundException.php
@@ -1,5 +1,5 @@
 <?php
-namespace Inachis\Component\CoreBundle;
+namespace Inachis\Component\CoreBundle\Exception;
 
 class FileNotFoundException extends \Exception
 {
