+++ src/Inachis/CoreBundle/Exception/InvalidEnvironmentException.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Inachis\Component\Common\Exception;
+namespace Inachis\Component\CoreBundle\Exception;
 
 class InvalidEnvironmentException extends \Exception
 {
